basic film app
Created basic structure of the flm application. Two pages - home and movies. On home on load page displays top 20 trending films. User can search film by title. When user is on the movies film the search is working, however if user is on the home page the search is no wroking after redirect yet