home page update
Updated the home page so that it has movies, tv series and people trending sections. Updated the request for movies search so that it doesn't include porn. updated design of home page. updated cards, eparate for movies series and people. updated placeholder component so that it receives the number of placeholders do display and renders that many placeholders.