film details page update
Updates the films details page. Displays comprehensive information about the film including cast and similar films. has minimal design. to be wroked on. Each movie card on main page is a link now to its film sdetails page. the people cards also are linking to their dedicated peorson's page. to be worked on. the genres for tv shows work properly now. the go back btn was updated in design.